feat: add component item-card
