+++ src/components/item-card/index.ts
@@ -1,0 +1,3 @@
+import ItemCard from './item-card';
+
+export default ItemCard;
