+++ src/components/item-list/item-list.css
@@ -3,27 +3,6 @@
   justify-content: space-evenly;
 }
 
-.character-card {
-  display: flex;
-  flex-direction: row;
-  width: 400px;
-  transition: all ease-in 0.3s;
-  cursor: pointer;
-}
-
-.character-card:hover {
-  border-color: rgb(55, 90, 127);
-}
-
-.character-image {
-  width: 150px;
-  flex-shrink: 0;
-  border-radius: 5px;
-  background-image: url('../../assets/icons/default_character.png');
-  background-size: cover;
-  background-position: center;
-}
-
 .section-left {
   flex-grow: 1;
   display: flex;
@@ -33,34 +12,4 @@
 
 .list-message {
   flex-grow: 1;
-}
-
-.dwarf {
-  background-image: url('../../assets/icons/dwarf.png');
-}
-
-.hobbit {
-  background-image: url('../../assets/icons/hobbit.png');
-}
-
-.human {
-  background-image: url('../../assets/icons/human.png');
-}
-
-.elf,
-.elves {
-  background-image: url('../../assets/icons/elf.png');
-}
-
-.ents {
-  background-image: url('../../assets/icons/ents.png');
-}
-
-.maiar {
-  background-image: url('../../assets/icons/wizard.png');
-}
-
-.orc,
-.orcs {
-  background-image: url('../../assets/icons/orc.png');
 }
